refactor: replace Fiedler metric with adaptive Lanczos algorithm
- Replace direct eigendecomposition with iterative Lanczos method for Fiedler value computation
- Add adaptive convergence detection that terminates early when λ₂ stabilizes
- Integrate optimized algorithm directly into production FiedlerValue metric class
- Remove benchmark test file after successful production integration
- Optimize device management keeping large matrices on GPU while computing eigenvalues on CPU
